+++ hook/index.ts
@@ -1,1 +1,2 @@
 export { default as useThrottling } from "./useThrottling";
+export { default as useInfiniteScroll } from "./useInfiniteScroll";
